server seperation
